Refactor out most functions into lib/repos.ts
